display bug fix

+++ Func.h
@@ -6,7 +6,7 @@
 void Calculate();
 void write24c02(void);
 void read24c02(void);
-int keyscan();
+uchar keyscan();
 void ParamProcess();
 void NumProcess(unsigned int keynum);
 #endif
